# Reduce allocations in recvMessage() by re-using a buffer