This is working but I am opening it to get feedback as I have not yet measured anything. The goal is to reduce allocations in hot paths by re-using a buffer.

Hopefully I will have time to do some measurements soon.