@fbernier 2019-12-15
I added some small changes along with some hackish benchmark machinery that I think should be added in another PR, in case you ever want to keep them.

I had to work around the fact that the master branch has no benchmark defined so I had to pull in the `driver_test.go` file to it in order to make the comparison. I only reproduced the `BasicQuery` test and benchmarked it. In theory, the larger the result and the bigger the improvements would be. Here are the results:

```
benchmark                 old ns/op     new ns/op     delta
BenchmarkBasicQuery-2     9201026       9093504       -1.17%

benchmark                 old allocs     new allocs     delta
BenchmarkBasicQuery-2     176            153            -13.07%

benchmark                 old bytes     new bytes     delta
BenchmarkBasicQuery-2     6881          6331          -7.99%
```