@huebnerr 2019-12-17
@fbernier, is this ready to be merged?