@fbernier 2019-12-17
There are some more things I am looking at but I think this is a good first step that could be merged in yes. I would not the travis changes and benchcmp script though. We could merge only the benchmark, although feel free to improve it, and add benchcmp in another PR where I can get rid of the driver_test.go checkout. I can take care of cleaning this up later if you want.