@fbernier 2019-12-19
Finally I think this is ready as-is. Since it's in a post-hook the benchmarks can never make a build fail anyway. There could be some cleanups to do in the future but having this in as-in right now sounds fine to me.