@fbernier 2020-01-10
@huebnerr I removed the benchmark as I currently don't have a clean way of having them in but I might revisit later. This is ready to merge IMO.